% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bothTests.R
\name{bothTests}
\alias{bothTests}
\title{Both, Traditionl (Chi square & Shapiro Wilk) and equivalence tests to detect normality}
\usage{
bothTests(x, mu = 0, sigma = 1, alpha = 0.05, epsilon = 0.4)
}
\arguments{
\item{x}{A numeric vector with sample data.}

\item{mu}{Theoretical mean from normal distribution.}

\item{sigma}{Theoretical standard deviation from normal distribution.}

\item{alpha}{Significance level of test.}

\item{epsilon}{Irrelevance/Equivalence limit.}
}
\value{
logical (TRUE if normality is detected, FALSE if not) value to, chi-square,
Shapiro - Wilk and equivalence normality test.
In addition, p-value (for chi square) and Upper
limit with \eqn{\epsilon^2} (for equivalence) are shown as attribute.
}
\description{
Chi square & Shapiro Wilk (goodnes of fit) and equivalence test (lack of fit) to detect
normality
}
\details{
Since we are fitting to a normal distribution with
parameters \eqn{\mu}  and \eqn{\sigma} known, chi-square statistic
uses k-1 degree of freedom. It is a fact only validate for
purposes of paper in reference, in real world parameters of a data set are
always unknown and chi-square statistics to prove normality uses k-3 degree
of freedom.

Normality is detected in chi-square test when p-value is greater than the
significance level \code{alpha} and for the equivalence test we declare
normality when upper limit of euclidean square distance is less than
irrelevance limit \eqn{\epsilon^2} \code{epsilon}^2
}
\examples{
# Proving if data comes from a normal standard
set.seed(123)
data <- rnorm(30)
bothTests(x = data)

# or from a normal N(mu = 5, sigma = 2) changing the default parameters
set.seed(123)
data <- rnorm(30, mean = 5, sd = 2)
bothTests(x = data, alpha = 0.01, epsilon = 0.15,
          mu = 5, sigma = 2)

}
