% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsilon.R
\name{epsilon}
\alias{epsilon}
\title{Non arbitrary Epsilon (Simulation Function)}
\usage{
epsilon(nSim, n1, n2, alpha = 0.05, seed = 123)
}
\arguments{
\item{nSim}{number of simulations.}

\item{n1}{Size of sample 1.}

\item{n2}{Size of sample 2.}

\item{alpha}{Significance level of test.}

\item{seed}{simulation seed}
}
\value{
A non arbitrary value of equivalence/irrelevance for a lack of fit test to
prove normality
}
\description{
Simulation function to compute a non arbitrary epsilon value for an
equivalence normality test of lack of fit.
}
\details{
epsilon resulting from this function should be used only for the purpose
of pretesting normality assumption when testing means difference
}
\examples{
eps <- epsilon(nSim = 10000, n1 = 7, n2 = 3)
eps

# If the epsilon value is adequate, then for a normal sample, the frequency
# with which an equivalence of normality test fails
# (it is not declared as normal) should be low.

nfails <- replicate(10000, {
  data <- rnorm(10)
  test <- normequiv(data, epsilon = eps)
  test["upper bound"] >= test["eps^2"]
})
sum(nfails) / 10000

}
