% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lackFitTest.R
\name{lackFitTest}
\alias{lackFitTest}
\title{Lack of fit test}
\usage{
lackFitTest(probs, afr, alpha = 0.05, epsilon = 0.4)
}
\arguments{
\item{probs}{Theoretical probabilities from a target distribution (for example: binomial, normal, etc).}

\item{afr}{Absolute frequencies obtained from a sample}

\item{alpha}{Significance level of test}

\item{epsilon}{Irrelevance/Equivalence limit}
}
\value{
a vector with Euclidean distance and upper bound of confidence interval.
Decision (Reject or not the null hypothesis) is showed as an attribute
}
\description{
Lack of fit test based on equivalence approach to detect if a sample comes
from a target distribution
}
\details{
In \eqn{H_0:d^2(\pi, \pi^0) \geq \epsilon^2} (lack of fit) Vs \eqn{H_1:d^2(\pi, \pi^0) < \epsilon^2},
null hypothesis \eqn{H_0} is rejected (proving fit to target distribution)
when upper bound of confidence interval is less than \eqn{\epsilon^2}.
}
\examples{
# Proving if x comes from a binomial distribution (n = 3, p = 0.45)
set.seed(123)
x <- rbinom(100, size = 3, prob = 0.45)
frequencies <- table(x)
binomProbs <- dbinom(0:3, size = 3, prob = 0.45)
lackFitTest(probs = binomProbs, afr = frequencies)

# or changing the default values
lackFitTest(probs = binomProbs, afr = frequencies,
            alpha = 0.01, epsilon = 0.15)

}
