% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normequiv.R
\name{normequiv}
\alias{normequiv}
\title{Lack of fit test to prove normality}
\usage{
normequiv(x, mu = 0, sigma = 1, alpha = 0.05, epsilon = 0.4)
}
\arguments{
\item{x}{a numeric vector with sample data.}

\item{mu}{theoretical mean from normal distribution.}

\item{sigma}{theoretical standard deviation from normal distribution.}

\item{alpha}{significance level of test.}

\item{epsilon}{irrelevance/equivalence limit.}
}
\value{
a vector with Euclidean distance, upper bound of confidence interval
and irrelevance limit \eqn{\epsilon^2}. Decision (Normality or non normality) over the
is showed as an attribute
}
\description{
Lack of fit test based on equivalence approach to prove normality
}
\details{
In \eqn{H_0:d^2(\pi, \pi^0) \geq \epsilon^2} (lack of fit) Vs
\eqn{H_1:d^2(\pi, \pi^0) < \epsilon^2}, null hypothesis \eqn{H_0} is
rejected (proving normality) when upper bound of confidence interval
is less than \eqn{\epsilon^2}
}
\examples{
# Proving if data comes from a normal standard
set.seed(123)
data <- rnorm(30)
normequiv(x = data)

# or from a normal N(mu = 5, sigma = 2) changing the default parameters
set.seed(123)
data <- rnorm(30, mean = 5, sd = 2)
normequiv(x = data, alpha = 0.01, epsilon = 0.15,
         mu = 5, sigma = 1)

}
