% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejectH0.R
\name{rejectH0}
\alias{rejectH0}
\title{H0 rejection proportion for all tests (Simulation Function).}
\usage{
rejectH0(
  nSim,
  n1,
  n2,
  epsilon,
  diffMean = 0,
  f.coef = c(0, 1, 0, 0),
  alpha = 0.05,
  seed = 123
)
}
\arguments{
\item{nSim}{number of simulations.}

\item{n1}{Size of sample 1.}

\item{n2}{Size of sample 2.}

\item{epsilon}{Irrelevance/Equivalence limit}

\item{diffMean}{Theoretical difference of means (= 0 by default)}

\item{f.coef}{Fleishman coefficients indicating normal contamination degree.}

\item{alpha}{Significance level of test.}

\item{seed}{simulation seed}
}
\value{
An estimation of Type I Error Probability TIEP (if diffMean = 0) or Power
(if diffMean > 0) of tests.
}
\description{
Simulation function to compute the proportion of rejections of H0 for four
comparisson mean tests: i) t-student direct (without pretest), ii) wilcoxon
direct (without pretest), iii) t-student or wilcoxon depending of chi-square
(goodnes of fit) pretest for normality, iv) t-student or wilcoxon depending
of Shapiro-Wilk (goodnes of fit) pretest for normality and v) t-student or
wilcoxon depending of equivalence (lack of fit) pretest for normality
}
\details{
when  \code{diffMean} is 0, function computes rejections proportions of
\eqn{H_0:\mu_1-\mu_2=0} when it is true i.e. TIEP,  while if \code{diffMean}
is different to zero function computes rejections proportions of
\eqn{H_0:\mu_1-\mu_2=0} when it is false i.e. Power.
}
\examples{
# TIEP in a non contamination sample (Normal):
rejectH0(nSim = 1000, n1 = 5, n2 = 3, epsilon = 0.4, diffMean = 0)

# TIEP in a non normal sample (contaminated)
rejectH0(nSim = 1000, n1 = 5, n2 = 3, epsilon = 0.4, diffMean = 0,
         f.coef = c(-0.3137, 0.8632, 0.3137, 0.0227))

# POWER in a non normal sample (contaminated)
rejectH0(nSim = 1000, n1 = 5, n2 = 3, epsilon = 0.4, diffMean = 3,
         f.coef = c(-0.3137, 0.8632, 0.3137, 0.0227))

}
