% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnonorm.R
\name{rnonorm}
\alias{rnonorm}
\title{Random non normal}
\usage{
rnonorm(n, f.coef)
}
\arguments{
\item{n}{number of observations.}

\item{f.coef}{Fleishman coefficients indicating normal contamination degree}
}
\value{
a vector of length \code{n} containing a non normal sample
}
\description{
Random number generation for a non normal standard distribution using contamination
degree based on Fleishman coefficients (altering skewness and kurtosis)
}
\details{
Being \eqn{X} a normal standard,  linear combination
\eqn{Z = a + bX + cX^2 + dX^3} follows an unknown distribution with skewness
\eqn{\gamma_1} and \eqn{\gamma_2} defining contamination degree from normal.
Coefficients \eqn{a, b, c, d} are known as Fleishman coefficients.
For example for  \eqn{a = 0, b = 1, c = 0, d = 0} does not exist contamination
}
\examples{
# No contaminated sample:
ncs <- rnonorm(n = 100, f.coef = c(0, 1, 0, 0))
hist(ncs, probability = TRUE)
curve(dnorm(x), add = TRUE, col = "red")

# Contaminated sample:
cs <- rnonorm(n = 100, f.coef = c(-0.3137, 0.8632, 0.3137, 0.0227))
hist(cs, probability = TRUE)
curve(dnorm(x), add = TRUE, col = "red")

}
