% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiepT.R
\name{tiepT}
\alias{tiepT}
\title{H0 rejection proportion  for t-student (Simulation Function)}
\usage{
tiepT(nSim, n1, n2, f.coef = c(0, 1, 0, 0), alpha = 0.05, seed = 123)
}
\arguments{
\item{nSim}{number of simulations.}

\item{n1}{Size of sample 1.}

\item{n2}{Size of sample 2.}

\item{f.coef}{Fleishman coefficients indicating normal contamination degree.}

\item{alpha}{Significance level of test.}

\item{seed}{simulation seed}
}
\value{
An estimation of Type I Error Probability TIEP for t-student test to
means comparison.
}
\description{
Simulation function to compute TIEP estimation in a t-student test for
means comparison with normal or non-normal samples
}
\examples{
# TIEP in a non contamination sample (Normal):
tiepT(nSim = 1000, n1 = 5, n2 = 3)

# TIEP in a non normal sample (contaminated)
tiepT(nSim = 1000, n1 = 5, n2 = 3,
     f.coef = c(-0.3137, 0.8632, 0.3137, 0.0227))

}
